requirements.txt
fastapi
uvicorn
pandas
mysql-connector-python
python-dotenv
ibm-watsonx-ai
scikit-learn